refactor: 기존 service 인터페이스와 구현체로 분리
